+++ encord/orm/storage.py
@@ -3,8 +3,7 @@ from enum import auto
 from typing import Dict, List, Optional, Union
 from uuid import UUID
 
-from orm.base_types import BeEnum, CamelStrEnum
-
+from encord.orm.analytics import BeEnum, CamelStrEnum
 from encord.orm.base_dto import BaseDTO, Field
 from encord.orm.dataset import LongPollingStatus
